+++ cli/lm-cmd.py
@@ -450,6 +450,8 @@ class LM_CLI(GenericCLI, ControllerCLI):
                 d['status'] = ['stopped', 'running', 'stopping'][d['status']]
             if api_func == 'list_controllers':
                 d['type'] = 'static' if d['static'] else 'dynamic'
+                if d.get('compress'):
+                    d['proto'] += '.z'
                 d['proto'] += '/' + ('mqtt' if d.get('mqtt_update') else 'ws')
             elif api_func == 'result':
                 import pytz