+++ cli/sfa-cmd.py
@@ -228,6 +228,8 @@ class SFA_CLI(GenericCLI, ControllerCLI, LECLI):
                                                    tz).isoformat()
             if api_func == 'list_controllers':
                 d['type'] = 'static' if d['static'] else 'dynamic'
+                if d.get('compress'):
+                    d['proto'] += '.z'
                 d['proto'] += '/' + ('mqtt' if d.get('mqtt_update') else 'ws')
             if api_func in ['list_macros', 'list_cycles', 'list_controllers']:
                 d['id'] = d['full_id']